LangChain/DocumentLoading/requirements.txt
openai
python-dotenv
langchain
datetime
wikipedia
langchain_experimental
langchain_openai
numexpr
pypdf
yt_dlp
pydub